requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.7.2
beautifulsoup4==4.12.0
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
distlib==0.3.6
Django==4.2.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
fastapi==0.115.7
filelock==3.9.0
Flask==2.2.2
fonttools==4.43.1
graphviz==0.20.3
greenlet==3.1.1
h11==0.14.0
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
Mako==1.3.8
MarkupSafe==2.1.2
matplotlib==3.8.0
networkx==3.4.1
numpy==1.24.2
packaging==23.2
Pillow==10.0.0
platformdirs==2.6.2
psutil==5.9.8
psycopg2==2.9.6
pydantic==2.10.6
pydantic_core==2.27.2
pygame==2.1.3
PyJWT==2.9.0
pyparsing==3.1.1
pyperclip==1.8.2
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
pywin32==306
PyYAML==6.0.2
requests==2.28.2
scipy==1.11.3
six==1.16.0
sniffio==1.3.1
Solution==5.4.0
soupsieve==2.4
SQLAlchemy==2.0.37
sqlparse==0.4.4
starlette==0.45.3
typing_extensions==4.12.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.14
uvicorn==0.34.0
virtualenv==20.17.1
Werkzeug==2.2.2
